Update README.md
News media application was designed on a motive to provide up-to-date news happening around the world. The application covers all field news such as Rural news, City news, State news, Country news, World news, Sports, Entertainment, Economics and Politics. The main focus of this application is to connect news articles from all around the world and deliver it to user as fast as possible in best visualize way. A news application is a big interactive database that tells a news story.